Update and rename Ninjas Garden.java to January-6.java
